introduce validate assignment
